+++ src/netprotocols/layer7/__init__.py
@@ -1,4 +1,19 @@
 from netprotocols.layer7.dhcp import DHCP
-from netprotocols.layer7.dns import DNS, DNSOverTCP, DNSResourceRecord
+from netprotocols.layer7.dns import (
+    DNS,
+    DNSOverTCP,
+    DNSQuestion,
+    DNSResourceRecord,
+    MXRecord,
+    SOARecord,
+)
 
-__all__ = ["DHCP", "DNS", "DNSOverTCP", "DNSResourceRecord"]
+__all__ = [
+    "DHCP",
+    "DNS",
+    "DNSOverTCP",
+    "DNSQuestion",
+    "DNSResourceRecord",
+    "MXRecord",
+    "SOARecord",
+]